Use simplified Program.cs - VetClinicPublic
